+++ src/main/scala/onion/compiler/SemanticError.scala
@@ -133,6 +133,7 @@ object SemanticError {
   case object EXAMPLE_FAILED extends SemanticError(65)
   case object RAW_TYPE_NOT_ALLOWED extends SemanticError(66)
   case object MISSING_RETURN extends SemanticError(67)
+  case object OVERRIDE_TARGET_NOT_FOUND extends SemanticError(68)
 }
 sealed abstract class SemanticError(val code: Int) {
   /** Returns the error code in format "E0001" */